+++ src/singleton.ts
@@ -2,15 +2,15 @@
  * Simple singleton mechanism
  */
 const _instances : any[] = [];
-export function get<T>(ctor: (new () => T)): T {
+export function get<T>(ctor: (new (...args: any[]) => T)): T {
     return getInstance(ctor);
 }
-export function register<T>(ctor: (new () => T), instance: T): T {
+export function register<T>(ctor: (new (...args: any[]) => T), instance: T): T {
     return registerInstance(ctor, instance);
 }
-export function getInstance<T>(ctor: (new () => T)): T {
+export function getInstance<T>(ctor: (new (...args: any[]) => T)): T {
     return _instances[ctor.name] || (_instances[ctor.name] = new ctor());
 }
-export function registerInstance<T>(ctor: (new () => T), instance: T): T {
+export function registerInstance<T>(ctor: (new (...args: any[]) => T), instance: T): T {
     return (_instances[ctor.name] = instance);
 }